+++ command/command_test.go
@@ -1,8 +1,6 @@
 package command
 
 import (
-	"flag"
-	"github.com/urfave/cli"
 	"strings"
 	"testing"
 )
@@ -36,28 +34,5 @@ func TestApp(t *testing.T) {
 				t.Error("expecting match but got error", err)
 			}
 		})
-	}
-}
-
-func TestNewTest(t *testing.T) {
-	testcommand := newTest(&Options{})
-	fn, ok := testcommand.Action.(func(c *cli.Context) error)
-
-	if !ok {
-		t.Error("cannot cast test command action to function")
-		return
-	}
-
-	set := flag.NewFlagSet("test", 0)
-	set.Parse([]string{"command_test.eskip"})
-	set.Bool("v", true, "")
-	set.String("p", "/bar", "")
-	ctx := cli.NewContext(nil, set, nil)
-	ctx.Command = testcommand
-
-	err := fn(ctx)
-
-	if err != nil {
-		t.Error(err)
 	}
 }
